+++ Scripts/HintDialog.cs
@@ -1,5 +1,5 @@
 using Godot;
-using System;
+using System.Threading.Tasks;
 using ArchipelagoMultiTextClient.Scripts;
 using CreepyUtil.Archipelago;
 
@@ -7,16 +7,34 @@ public partial class HintDialog : ConfirmationDialog
 {
     public ApClient Client;
     private string _Item;
+    private string _Location;
+    private bool _IsLocation;
 
     public string Item
     {
         get => _Item;
-        set => DialogText = $"Hint for item:\n{_Item = value}?";
+        set
+        {
+            _IsLocation = false;
+            DialogText = $"Hint for item:\n{_Item = value}?";
+        }
     }
 
-    public override void _Ready() => Confirmed += () =>
+    public string Location
     {
-        Client.Say($"!hint {_Item}");
-        MainController.MoveToTab = 1;
-    };
+        get => _Location;
+        set
+        {
+            _IsLocation = true;
+            DialogText = $"Hint locations:\n{_Location = value}?";
+        }
+    }
+
+    public override void _Ready()
+        => Confirmed += () =>
+        {
+            Task.Delay(500).GetAwaiter().GetResult();
+            Client.Say(_IsLocation ? $"!hint_location {_Location}" : $"!hint {_Item}");
+            MainController.MoveToTab = 1;
+        };
 }