+++ Scripts/SlotClient.cs
@@ -6,12 +6,13 @@ using Archipelago.MultiClient.Net.Enums;
 using Archipelago.MultiClient.Net.Packets;
 using ArchipelagoMultiTextClient.Scripts;
 using CreepyUtil.Archipelago;
-using static TextClient;
+using static ArchipelagoMultiTextClient.Scripts.MainController;
 
 public partial class SlotClient : PanelContainer
 {
+    private static int ClientCount = 0;
     public bool IsRunning;
-    
+
     [Export] private Button _ConnectButton;
     [Export] private Label _ConnectingLabel;
     [Export] private Button _DisconnectButton;
@@ -19,8 +20,9 @@ public partial class SlotClient : PanelContainer
     [Export] private Button _DeleteButton;
     [Export] private RichTextLabel _ErrorLabel;
     public ApClient Client = new();
+    public bool IsTextClient = false;
     private MainController _Main;
-    
+
     public string PlayerName { get; private set; }
 
     public void Init(MainController main, string playerName)
@@ -39,45 +41,62 @@ public partial class SlotClient : PanelContainer
 
     public void TryConnection()
     {
+        if (ClientCount >= 7)
+        {
+            ConnectionFailed(["Can only have 1 slots connected"]);
+            return;
+        }
+        
+        if (ConnectionCooldown > 0)
+        {
+            ConnectionFailed(["Please wait 15s after connecting/changing slots to do so again"]);
+            return;
+        }
+        
         IsRunning = true;
         _ErrorLabel.Visible = false;
         _ConnectButton.Visible = false;
         _DeleteButton.Visible = false;
         _ConnectingLabel.Visible = true;
         LoginInfo login = new(_Main.Port, PlayerName, _Main.Address, _Main.Password);
+        
+        ClientCount++;
+        ConnectionCooldown = 15;
 
+        string[] tags = ChosenTextClient is null ? ["TextOnly"] : ["TextOnly", "NoText"];
+        
         Task.Run(() =>
         {
             try
             {
-                string[]? error = null;
+                string[] error = null;
                 lock (Client)
                 {
-                    error = Client.TryConnect(login, 0, "", ItemsHandlingFlags.NoItems, tags: ["TextOnly"]);
+                    error = Client.TryConnect(login, 0, "", ItemsHandlingFlags.NoItems, tags: tags);
                     Client.Session.Socket.PacketReceived += packet =>
                     {
                         switch (packet)
                         {
                             case ChatPrintJsonPacket message:
-                                AwaitingMessages.Enqueue(new ClientMessage(message.Data, chatPrintJsonPacket: message));
+                                TextClient.Messages.Enqueue(new ClientMessage(message.Data, chatPrintJsonPacket: message));
                                 break;
                             case BouncedPacket:
                                 break;
                             case PrintJsonPacket updatePacket:
                                 if (updatePacket.Data.Length == 1)
                                 {
-                                    AwaitingMessages.Enqueue(new ClientMessage(updatePacket.Data, isServer: true));
+                                    TextClient.Messages.Enqueue(new ClientMessage(updatePacket.Data, isServer: true));
                                 }
 
                                 if (updatePacket.Data.Length < 2) break;
                                 if (updatePacket.Data.First().Text!.StartsWith("[Hint]: "))
                                 {
                                     if (updatePacket.Data.Last().HintStatus!.Value == HintStatus.Found) break;
-                                    AwaitingMessages.Enqueue(new ClientMessage(updatePacket.Data));
+                                    TextClient.Messages.Enqueue(new ClientMessage(updatePacket.Data));
                                 }
                                 else if (updatePacket.Data[1].Text is " found their " or " sent ")
                                 {
-                                    AwaitingMessages.Enqueue(new ClientMessage(updatePacket.Data, true));
+                                    TextClient.Messages.Enqueue(new ClientMessage(updatePacket.Data, true));
                                 }
 
                                 break;
@@ -132,12 +151,14 @@ public partial class SlotClient : PanelContainer
     public void HasDisconnected()
     {
         IsRunning = false;
+        _ErrorLabel.Visible = false;
         _ConnectingLabel.Visible = false;
         _ConnectButton.Visible = true;
         _DeleteButton.Visible = true;
         _DisconnectButton.Visible = false;
         _Main.DisconnectClient(Client);
         Client = new ApClient();
+        ClientCount--;
     }