+++ Scripts/TextClient.cs
@@ -1,11 +1,10 @@
 using System;
-using Godot;
 using System.Collections.Concurrent;
+using Godot;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Text;
-using Archipelago.MultiClient.Net.Converters;
+using System.Threading.Tasks;
 using Archipelago.MultiClient.Net.Enums;
 using Archipelago.MultiClient.Net.Models;
 using Archipelago.MultiClient.Net.Packets;
@@ -14,9 +13,12 @@ using static ArchipelagoMultiTextClient.Scripts.MainController;
 
 public partial class TextClient : VBoxContainer
 {
+    public static ConcurrentQueue<ClientMessage> Messages = [];
+    public static bool HintRequest;
+    public static double HintRequestTimer;
     public static bool RefreshText;
     public static bool ClearClient;
-    public static ConcurrentQueue<ClientMessage> AwaitingMessages = [];
+    public static List<string> CopyList = [];
 
     [Export] private MainController _Main;
     [Export] private OptionButton _SelectedClient;
@@ -25,26 +27,22 @@ public partial class TextClient : VBoxContainer
     [Export] private RichTextLabel _Messages;
     [Export] private LineEdit _SendMessage;
     [Export] private Button _SendMessageButton;
-    [Export] private SpinBox _Latency;
     [Export] private Button _ScrollToBottom;
     [Export] private ScrollContainer _ScrollContainer;
     private Queue<ClientMessage> _ChatMessages = new(250);
     private Queue<ClientMessage> _ItemLog = new(250);
     private Queue<ClientMessage> _Both = new(500);
     private List<string> _SentMessages = new(50);
-    private MessageGateKeeper _Keeper = new();
     private ScrollBar _VScrollBar;
     private string[] _CurrentPlayers = [];
     private int _ScrollBackNum;
     private bool _ToScroll;
     private double _MessageCooldown;
+    private long _LastSelected;
 
     public override void _Ready()
     {
         _VScrollBar = _ScrollContainer.GetVScrollBar();
-
-        _Latency.ValueChanged += val => MainController.Data.Latency = MessageGateKeeper.Latency = (long)val;
-        _Latency.Value = MessageGateKeeper.Latency = MainController.Data.Latency;
 
         _WordWrap.ItemSelected += i
             => _Messages.AutowrapMode = (TextServer.AutowrapMode)(MainController.Data.WordWrap = i);
@@ -78,7 +76,7 @@ public partial class TextClient : VBoxContainer
             _ScrollBackNum = Math.Clamp(_ScrollBackNum, 0, _SentMessages.Count - 1);
             _SendMessage.Text = _SentMessages[_ScrollBackNum];
         };
-        _SendMessage.FocusEntered += () => _ScrollBackNum = 0;
+        _SendMessage.FocusEntered += () => _ScrollBackNum = _SentMessages.Count - 1;
 
         _SendMessageButton.Pressed += () => SendMessage(_SendMessage.Text);
         _ScrollToBottom.Pressed +=
@@ -89,21 +87,71 @@ public partial class TextClient : VBoxContainer
             _ScrollContainer.ScrollVertical = (int)_ScrollContainer.GetVScrollBar().MaxValue;
             _ToScroll = false;
         };
+
+        _SelectedClient.ItemSelected += l =>
+        {
+            if (ConnectionCooldown > 0)
+            {
+                _SelectedClient.Selected = (int)_LastSelected;
+                return;
+            }
+
+            ChosenTextClient.Session.ConnectionInfo.UpdateConnectionOptions(["TextOnly", "NoText"]);
+            Task.Delay(750).GetAwaiter().GetResult();
+            ChosenTextClient = null;
+            ChosenTextClient = ActiveClients[(int)l];
+            ChosenTextClient.Session.ConnectionInfo.UpdateConnectionOptions(["TextOnly"]);
+            ConnectionCooldown = 15;
+            _LastSelected = l;
+        };
+
+        _Messages.MetaClicked += s => DisplayServer.ClipboardSet(CopyList[int.Parse((string)s)]);
     }
 
     public override void _Process(double delta)
     {
         _SendMessageButton.Disabled = _MessageCooldown > 0;
-        
+
         if (_MessageCooldown > 0)
         {
             _MessageCooldown -= delta;
         }
-        
+
         if (ClearClient)
         {
             Clear();
             ClearClient = false;
+        }
+
+        if (HintRequest)
+        {
+            HintRequestTimer += delta;
+        }
+
+        if (!HintRequest && !Messages.IsEmpty)
+        {
+            while (!HintRequest && !Messages.IsEmpty)
+            {
+                Messages.TryDequeue(out var message);
+                _ToScroll = _VScrollBar.Value >= _VScrollBar.MaxValue - _ScrollContainer.Size.Y;
+                _Both.Enqueue(message);
+
+                if (message.IsItemLog)
+                {
+                    _ItemLog.Enqueue(message);
+                }
+                else
+                {
+                    _ChatMessages.Enqueue(message);
+                }
+
+                if (message.IsHintRequest)
+                {
+                    HintRequest = true;
+                }
+            }
+
+            RefreshText = true;
         }
 
         if (_CurrentPlayers.Length != ActiveClients.Count)
@@ -116,27 +164,9 @@ public partial class TextClient : VBoxContainer
             }
         }
 
-        while (!AwaitingMessages.IsEmpty)
-        {
-            _ToScroll = _VScrollBar.Value >= _VScrollBar.MaxValue - _ScrollContainer.Size.Y;
-            AwaitingMessages.TryDequeue(out var message);
-            if (!_Keeper.CanAdd(message.Message)) continue;
-            _Both.Enqueue(message);
-
-            if (message.IsItemLog)
-            {
-                _ItemLog.Enqueue(message);
-            }
-            else
-            {
-                _ChatMessages.Enqueue(message);
-            }
-
-            RefreshText = true;
-        }
-
         if (!RefreshText) return;
 
+        CopyList.Clear();
         switch (_Content.Selected)
         {
             case 0: // text only
@@ -160,10 +190,10 @@ public partial class TextClient : VBoxContainer
         if (trim == "") return;
         if (_CurrentPlayers.Length == 0) return;
         _MessageCooldown = .3f;
-        _ScrollBackNum = 0;
         _SentMessages.Add(trim);
-        ActiveClients[_SelectedClient.Selected].Say(message);
+        ChosenTextClient!.Say(message);
         _SendMessage.Text = "";
+        _ScrollBackNum = _SentMessages.Count - 1;
     }
 
     public void Clear()
@@ -172,9 +202,8 @@ public partial class TextClient : VBoxContainer
         _ChatMessages.Clear();
         _ItemLog.Clear();
         _Both.Clear();
-        AwaitingMessages.Clear();
-        _Keeper.PacketMessageQueue.Clear();
-        _Keeper.MessageSet.Clear();
+        Messages.Clear();
+        HintRequest = false;
     }
 }
 
@@ -182,28 +211,33 @@ public readonly struct ClientMessage(
     JsonMessagePart[] messageParts,
     bool isItemLog = false,
     bool isServer = false,
-    ChatPrintJsonPacket? chatPrintJsonPacket = null)
+    ChatPrintJsonPacket chatPrintJsonPacket = null)
 {
     public readonly bool IsItemLog = isItemLog;
     public readonly bool IsServer = isServer;
     public readonly JsonMessagePart[] MessageParts = messageParts;
-    public readonly string Message = string.Join("", messageParts.Select(msg => msg.Text));
-    public readonly ChatPrintJsonPacket? ChatPacket = chatPrintJsonPacket;
+    public readonly ChatPrintJsonPacket ChatPacket = chatPrintJsonPacket;
+
+    public readonly bool IsHintRequest =
+        chatPrintJsonPacket is not null && chatPrintJsonPacket.Message.StartsWith("!hint");
 
     public string GenString()
     {
         string color;
+        var copyId = TextClient.CopyList.Count;
         if (ChatPacket is not null)
         {
             color = PlayerColor(ChatPacket.Slot);
-            return $"[color={color}]{GetAlias(ChatPacket.Slot, true)}[/color]: {ChatPacket.Message.Clean()}";
+            TextClient.CopyList.Add($"{GetAlias(ChatPacket.Slot, false)}: {ChatPacket.Message}");
+            return $"[color={color}][url={copyId}]{GetAlias(ChatPacket.Slot, true)}[/url][/color]: {ChatPacket.Message.Clean()}";
         }
 
         StringBuilder messageBuilder = new();
 
         if (IsServer)
         {
-            messageBuilder.Append($"[color={MainController.Data.ColorSettings["player_server"].Hex}]Server[/color]: ");
+            messageBuilder.Append(
+                $"[color={MainController.Data["player_server"].Hex}][url={copyId}]Server[/url][/color]: ");
         }
 
         foreach (var part in MessageParts)
@@ -222,55 +256,79 @@ public readonly struct ClientMessage(
                     break;
                 case JsonMessagePartType.LocationId:
                     var location = LocationIdToLocationName(long.Parse(part.Text), part.Player!.Value);
-                    color = MainController.Data.ColorSettings["location"];
+                    color = MainController.Data["location"];
                     messageBuilder.Append($"[color={color}]{location.Clean()}[/color]");
                     break;
                 case JsonMessagePartType.EntranceName:
                     var entranceName = part.Text.Trim();
-                    color = MainController.Data.ColorSettings[entranceName == "" ? "entrance_vanilla" : "entrance"];
+                    color = MainController.Data[entranceName == "" ? "entrance_vanilla" : "entrance"];
                     messageBuilder.Append(
                         $"[color={color}]{(entranceName == "" ? "Vanilla" : entranceName).Clean()}[/color]");
                     break;
                 case JsonMessagePartType.HintStatus:
                     var name = HintStatusText[(HintStatus)part.HintStatus!];
-                    color = MainController.Data.ColorSettings[HintStatusColor[(HintStatus)part.HintStatus!]];
+                    color = MainController.Data[HintStatusColor[(HintStatus)part.HintStatus!]];
                     messageBuilder.Append($"[color={color}]{name.Clean()}[/color]");
                     break;
                 default:
-                    messageBuilder.Append((part.Text ?? "").Clean());
+                    var text = (part.Text ?? "").Clean();
+                    messageBuilder.Append(text);
+                    if (IsServer)
+                    {
+                        TextClient.CopyList.Add(text);
+                    }
+
                     break;
             }
         }
 
         return messageBuilder.ToString();
     }
-}
 
-public class MessageGateKeeper
-{
-    public static long Latency = 50;
-    public PriorityQueue<DataTimestamp, long> PacketMessageQueue = new();
-    public HashSet<string> MessageSet = [];
+    public static readonly JsonMessagePart[] TextParts =
+    [
+        new() { Text = "[Hint]: " }, new() { Text = "'s " }, new() { Text = " is at " }, new() { Text = " in " },
+        new() { Text = "'s World. " },
+    ];
 
-    public bool CanAdd(string message)
+    public static implicit operator ClientMessage(Hint hint)
     {
-        var currentTimestamp = DateTimeOffset.Now.ToUnixTimeMilliseconds();
-        while (PacketMessageQueue.Count != 0 && currentTimestamp - PacketMessageQueue.Peek().Timestamp > Latency)
-        {
-            var dequeued = PacketMessageQueue.Dequeue();
-            MessageSet.Remove(dequeued.Message);
-        }
-
-        if (!MessageSet.Add(message)) return false;
-        PacketMessageQueue.Enqueue(new DataTimestamp(currentTimestamp, message), currentTimestamp);
-        return true;
+        return new ClientMessage([
+            TextParts[0],
+            new JsonMessagePart
+            {
+                Type = JsonMessagePartType.PlayerId,
+                Text = $"{hint.ReceivingPlayer}"
+            },
+            TextParts[1],
+            new JsonMessagePart
+            {
+                Type = JsonMessagePartType.ItemId,
+                Text = $"{hint.ItemId}",
+                Flags = hint.ItemFlags,
+                Player = hint.ReceivingPlayer
+            },
+            TextParts[2],
+            new JsonMessagePart
+            {
+                Type = JsonMessagePartType.LocationId,
+                Text = $"{hint.LocationId}",
+                Player = hint.FindingPlayer
+            },
+            TextParts[3],
+            new JsonMessagePart
+            {
+                Type = JsonMessagePartType.PlayerId,
+                Text = $"{hint.FindingPlayer}"
+            },
+            TextParts[4],
+            new JsonMessagePart
+            {
+                Type = JsonMessagePartType.HintStatus,
+                HintStatus = hint.Status
+            }
+        ]);
     }
-}
-
-public readonly struct DataTimestamp(long timestamp, string message)
-{
-    public readonly long Timestamp = timestamp;
-    public readonly string Message = message;
 }
 
 public static class TextHelper